Remove Citations
This PR proposes removing the 'citation' property entirely from the
specification. This is motivated by the following

* There is no *functional* difference between a citation and an
attestation. By this I mean that no change in the modelling is required
beyond changing the name of the property
* The specification uses highly academic and vague descriptions of
citations and there are no examples of citations. This would make this
property impractical to users of the model
* The distinction between citation and attestation does not seem to be
supported by literature other than by authors of this spec:
https://aclanthology.org/2020.globalex-1.1.pdf . This paper justifies a
citation as a "special form of citation that provides evidence of a
lexical phenomenom". This seems to be identical to the definition of
Tiberius et al. "A citation is a short extract from a text which
provides evidence for a word, phrase, usage, or meaning in authentic
use"
[source](https://campus.dariah.eu/resources/hosted/introduction-to-corpus-based-lexicographic-practice#lexicographic-evidence-from-slips-to-corpora)
* Removing Citation would fix #30, #31 and #46 immediately